(#6) Implementando padrão de projeto controlador
